style: add styling for about me sections
These styles focus on desktop and tablet users, later will add phone users styling.

fix: remove unnecessary sections in about me

To reach the desired outcome of the layouts I needed to use float so that the text could wrap alongside the image, but earlier we had containers that wouldn't have allowed for that.